# backend/ent/schema/subject.go
package schema

import (
	"github.com/facebookincubator/ent"
	"github.com/facebookincubator/ent/schema/edge"
	"github.com/facebookincubator/ent/schema/field"
)

// Subject holds the schema definition for the Subject entity.
type Subject struct {
	ent.Schema
}

// Fields of the Subject.
func (Subject) Fields() []ent.Field {
	return []ent.Field{
		field.String("subject_name").
			NotEmpty().
			Unique(),
	}
}

// Edges of the Subject.
func (Subject) Edges() []ent.Edge {
	return []ent.Edge{
		edge.From("owner",Teacher.Type).
			Ref("subjects").
			Unique(),
		edge.To("course_items", CourseItem.Type).
			StorageKey(edge.Column("subject_id")),
	}
}